Reestructuración de cuentas de usuarios. Actualización Laravel 5.5.

+++ database/migrations/2017_04_02_215101_create_transactions_table.php
@@ -18,7 +18,7 @@ class CreateTransactionsTable extends Migration
             $table->increments('id');
             $table->integer('sucursal_id');
             $table->decimal('amount')->default(0);
-            $table->integer('type')->default(0);
+            $table->smallInteger('type')->default(0);
             $table->date('date')->default(DB::raw('CURRENT_TIMESTAMP'));
             $table->time('time')->default(DB::raw('CURRENT_TIMESTAMP'));
             $table->foreign('sucursal_id')->references('id')->on('sucursales');
